windoze

+++ test/simul_1000_grains/run_simul.py
@@ -12,8 +12,10 @@ def measure(cmd):
     else:
         open("timing.log","w+").write("%f %s\n"%(end-start, cmd))
 
-os.remove("allgrid.map")
-os.remove("timing.log")    
+if os.path.exists("allgrid.map"):
+    os.remove("allgrid.map")
+if os.path.exists("timing.log"):
+    os.remove("timing.log")    
 measure("python make_u_t.py")
 measure("PolyXSim.py -i Al1000.inp")
 measure("python grid_index.py  Al1000/Al1000.flt Al1000/Al1000.par grid")
